hard mode training for control
Switched to a more proper UNIX keyboard layout (that key between tab,
shift, and 'A' is a control key now). But because I'm still a vim-er, I
like having escape so close to me. So, using xcape to make left control
be escape if I don't chord another key). But, decades of muscle memory
kept me hitting the corner control key. So now I have to hard mode for a
while.